% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts.R
\docType{data}
\name{abstracts}
\alias{abstracts}
\title{Sample Dataset: abstracts}
\format{
A data frame with 20 rows and 16 variables:
\describe{
  \item{title}{Title of the manuscript}
  \item{authors}{Authors}
  \item{abstract}{The column that contains all abstracts}
}
}
\source{
Created for demonstration purposes
}
\usage{
abstracts
}
\description{
This dataset includes sample abstracts downloaded from one of my Covidence project.
Pre-processing was done by running the clean_name function from janitor package.
}
\keyword{datasets}
