% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{add_predefined_vars}
\alias{add_predefined_vars}
\title{Pre-processing functions before shinyapps}
\usage{
add_predefined_vars(abstract_data, list_vars)
}
\arguments{
\item{abstract_data}{a csv file contains abstract information. This could be the csv file downloaded from covidence}

\item{list_vars}{a vector of data elements you want to extract. i.e. c("no_participants,"no_female","..")}
}
\value{
new dataset with additional empty columns
}
\description{
Pre-processing functions and saving intermediate data
}
