% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_manual_extractor.R
\name{glance_manual_extractor}
\alias{glance_manual_extractor}
\title{glance_manual_extractor}
\usage{
glance_manual_extractor()
}
\description{
A shinyapp that display the abstract and allow for randomly reading in data to glace what are the avaiable information in abstract.
}
