% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_manual_check.R
\name{llm_manual_check}
\alias{llm_manual_check}
\title{Launch the first shinyapp for to look into what variables are avaiable in the abstract.}
\usage{
llm_manual_check()
}
\value{
Shiny app.
}
\description{
This function calls a shinyapp that display the abstract and allow for randomly reading in data to glace what are the avaiable information in abstract.
}
