% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_validation_app.R
\name{manual_validation_app}
\alias{manual_validation_app}
\title{manual_validation_app}
\usage{
manual_validation_app()
}
\description{
This Shiny app allows you to manually review the results generated by the `process_with_ollama()` function. By uploading a testing dataset with LLM-extracted values, you can systematically check and validate each field filled out by the model.
}
\examples{

\dontrun{
manual_validation_app()
}

}
