% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaextractoR-package.R
\docType{package}
\name{metaextractoR-package}
\alias{metaextractoR}
\alias{metaextractoR-package}
\title{metaextractoR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Danyang Dai \email{danyang.dai@uq.edu.au}

}
\keyword{internal}
