% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering.R
\name{prompt_engineering}
\alias{prompt_engineering}
\title{prompt_engineering}
\usage{
prompt_engineering()
}
\description{
A shinyapp that designed for model testing and prompt engineering.
}
\examples{
\donotrun {
prompt_engineering()
}
}
