% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{save_stage_1_data}
\alias{save_stage_1_data}
\title{save_stage_1_data}
\usage{
save_stage_1_data(abstract_data)
}
\arguments{
\item{abstract_data}{a csv file contains abstract information as well as the variables you want to extract.}
}
\value{
a csv file saved in the metaextroctor_process_data file named stage_0_data.csv
}
\description{
This function will save the abstract data with empty columns to the processed data folder.
This csv file will be used in the shinyapp 1
The data will be stored in metaextroctor_process_data
}
