% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{save_training_data}
\alias{save_training_data}
\title{save_training_data}
\usage{
save_training_data(training_abs)
}
\arguments{
\item{training_abs}{training abstracts including the variables you want to extract.}
}
\value{
a csv file saved in the metaextroctor_process_data file named training_stage_0_data.csv
}
\description{
This function will save the training abstract data with empty columns to the processed data folder.
This csv file will be used in the shinyapp 1
The data will be stored in metaextroctor_process_data
}
