% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{separate_training}
\alias{separate_training}
\title{separate_training}
\usage{
separate_training(data, percentage = 0.1)
}
\arguments{
\item{data}{The csv file contains abstracts with}

\item{percentage}{percentage of separation training sets. If percentage set to be 0.1, 10% of the data will be the training set and 90% of the data will be testing set.}
}
\value{
A list with two the train and test data.frames.
}
\description{
This function will separate the abstracts into training and testing sets.
}
\examples{
separate_training(abstracts, percentage = 0.4)

}
