% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abstracts}
\alias{abstracts}
\alias{app_2}
\alias{app_3}
\alias{testing_stage_0_data}
\alias{training_stage_0_data}
\title{Sample abstracts}
\format{
A data frame with 20 rows and 16 variables:
\describe{
  \item{title}{Title of the manuscript}
  \item{authors}{Authors}
  \item{abstract}{The column that contains all abstracts}
  \item{published_year}{Year of the publication}
  \item{published_month}{Month of the publication}
  \item{journal}{Journal's name}
  \item{volume}{Volume number}
  \item{issue}{Issue number}
  \item{pages}{Page number}
  \item{accession_number}{Accession number}
  \item{doi}{doi number}
  \item{ref}{Reference}
  \item{covidence_number}{Covidence number}
  \item{study}{Study}
  \item{notes}{Notes}
  \item{tags}{Tags}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 15 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 40 rows and 16 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 40 rows and 16 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 16 columns.
}
\source{
Created for demonstration purposes
}
\usage{
abstracts

app_2

app_3

testing_stage_0_data

training_stage_0_data
}
\description{
This dataset includes sample abstracts downloaded from one of my Covidence project.
Pre-processing was done by running the clean_name function from janitor package.
}
\keyword{datasets}
