% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_manual_app.R
\name{glance_manual_app}
\alias{glance_manual_app}
\title{Launch the first shinyapp for to look into what variables are available in the abstract.}
\usage{
glance_manual_app()
}
\value{
Runs a shiny app.
}
\description{
This function launches a Shiny app that displays each abstract alongside the data elements you want to extract. The app provides an interactive interface for manually entering this information, helping to fill out the _manual column generated by the `add_predefined_vars()` function in the training set.
}
\examples{
if (interactive()) {
 glance_manual_app()
}
}
