% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_validation_app.R
\name{manual_validation_app}
\alias{manual_validation_app}
\title{manual_validation_app}
\usage{
manual_validation_app()
}
\description{
A shinyapp that manually check the LLM outputs
}
\examples{

\donttest{
manual_validation_app()
}

}
