% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_engineering_app.R
\name{prompt_engineering_app}
\alias{prompt_engineering_app}
\title{prompt_engineering_app}
\usage{
prompt_engineering_app()
}
\description{
A shinyapp that designed for model testing and prompt engineering.
}
\examples{

\donttest{
prompt_engineering_app()
}

}
