% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{separate_training}
\alias{separate_training}
\title{separate_training}
\usage{
separate_training(abstracts, percentage = 0.1, seed = 1111)
}
\arguments{
\item{abstracts}{The csv file contains abstracts with}

\item{percentage}{percentage of separation training sets.}
}
\description{
This function will separate the abstracts into training and testing sets.
}
